requirements.txt
flask
scikit-learn
pandas
numpy
pyserial
requests
joblib
matplotlib
